Work in progress
This commit begins the process of getting the codebase up to date.

Work has focused on import of Qualtrics data for long term storage on
ttracker. Admin pages (admin_index.php, adminajax.php, admin.css) in
particular, and more general library functions are in this update.

This does not bring things up to date.